Remove the `[[nodiscard]]` attrib for better readability
This attribute only serves as a warning for users to consume the
returned value. Note however that writing `[[nodiscard]]` before every
single function whose return value shouldn't be ignored makes the code
quite less pleasant to read and work with. Let's keep it simple and
remove this unnecessary attribute and stop treating the API consumers
like fools that need a guiding hand to program.

Signed-off-by: Luiz G. Mugnaini A. <luizmugnaini@gmail.com>